+++ edison-status/src/test/java/de/otto/edison/acceptance/api/StatusApi.java
@@ -25,6 +25,11 @@ public class StatusApi extends SpringTestBase {
     private static String content = null;
     private static HttpStatus statusCode;
 
+    public static When internal_is_retrieved_as(final String mediaType) throws IOException {
+        getResource("http://localhost:8085/teststatus/internal", of(mediaType));
+        return When.INSTANCE;
+    }
+
     public static When internal_status_is_retrieved_as(final String mediaType) throws IOException {
         getResource("http://localhost:8085/teststatus/internal/status", of(mediaType));
         return When.INSTANCE;